+++ src/test/java/subway/line/adapter/input/web/integrated/DeleteLineIntegratedTest.java
@@ -1,4 +1,4 @@
-package subway.line.adapter.input.web;
+package subway.line.adapter.input.web.integrated;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
